removed some warnings and minor fixes

+++ ff/distributed/ff_dprinter.hpp
@@ -16,7 +16,7 @@ namespace ff {
             if (c != std::char_traits<char>::eof()) {
                 if (this->need_prefix
                     && !this->prefix.empty()
-                    && this->prefix.size() != this->sbuf->sputn(&this->prefix[0], this->prefix.size())) {
+                    && (int)(this->prefix.size()) != this->sbuf->sputn(&this->prefix[0], this->prefix.size())) {
                     return std::char_traits<char>::eof();
                 }
                 this->need_prefix = c == '\n';
